+++ src/migrations/m260621_000000_add_audit_runs_table.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace kooba\contentaudit\migrations;
+namespace iistudio\contentaudit\migrations;
 
 use craft\db\Migration;